splatch/requirements.txt
# Python 3.13.2
numpy==2.3.4
Pillow==12.0.0
scikit_learn==1.7.2
timm==1.0.22
torch==2.9.1
torchvision==0.24.1
